refine graph_db_handler
Committed-by: xiaolei.zl from Dev container

Committed-by: xiaolei.zl from Dev container
